Fixing two minor display issue
When showing the full indexedDB entries, January was displayed as month 0 which is normal for JavaScript but not normal for humans
When clocking in, out and then in again, then the weektime counter would be initialized twice and the displayed value become doubled. 
Fixed by resetting the coutner. This bug was introduced by a refactoring just before putting this on Github